Kill child processes more effectively.
Okay, this is a little shaky, but here's the idea: when doing things like
`yerna run foo`, if `foo` spawned child processes of its own, Yerna is
not able to see that. It successfully kills its immediate children, but
the grandchildren (etc.) keep chugging along. We use a cute hack from
http://azimi.me/2014/12/31/kill-child_process-node-js.html (and also
suggested by a node maintainer a long time ago at
https://github.com/nodejs/node-v0.x-archive/issues/1811#issuecomment-2259986)
to make sure that we kill all of this processes' dependents. The downside
is that rogue children are more likely with severe failures, because the
OS won't auto-kill them when the parent exits.

This change is part of the "always use exitCode instead of exit" work,
otherwise we could just violently process.exit(1) and be done with it
(which I'm not opposed to, but avoiding exit helps track down latent
issues like the exact one that this commit fixes).